Fix build errors: enum casts, method naming, missing properties, and update CLAUDE.md
- Add explicit enum↔int casts in Conversation/Offer Mapper files
- Fix ValidateBrandExist→ValidateBrandExists, ValidateProductExist→ValidateProductExists, ValidateSellerExist→ValidateSellerExists
- Remove invalid 'using LivestockTrading.Domain.Enums' from Banner/TaxRate Mappers
- Add BannerErrors.StartDateRequired/EndDateRequired to DomainErrors
- Fix Seller.Name→Seller.BusinessName in Products Detail Mapper
- Update CLAUDE.md with enum casting rules and entity property reference

+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/Conversations/Queries/All/Mapper.cs
@@ -13,7 +13,7 @@ public class Mapper
 			ParticipantUserId2 = c.ParticipantUserId2,
 			ProductId = c.ProductId,
 			Subject = c.Subject,
-			Status = c.Status,
+			Status = (int)c.Status,
 			LastMessageAt = c.LastMessageAt,
 			CreatedAt = c.CreatedAt
 		}).ToList();
